requirements.txt
streamlit
numpy
tensorflow
opencv-python
pillow
streamlit-option-menu
pandas
scikit-learn
streamlit-lottie
matplotlib
pydot